Implement planning and execution management in HybridAgent
- Introduced ExecutionPlan and PlanStep classes to manage execution plans, enhancing modularity and clarity.
- Integrated planning logic into HybridAgent, allowing for dynamic creation and replanning of execution steps based on user goals and context.
- Removed redundant state management code, streamlining the architecture for better maintainability.
- Added support for saving and loading plans in both JSON and Markdown formats, improving usability and documentation.
- Updated dependencies in the project configuration to include aiofiles and structlog for enhanced functionality and logging capabilities.
